Adds ability to broacast different payloads.
Clients running different protocol version should be able to receive
different payloads.

+++ plugin/AndroidRemote/Commands/Requests/RequestProtocol.cs
@@ -15,8 +15,8 @@ namespace MusicBeePlugin.AndroidRemote.Commands.Requests
 
         public void Execute(IEvent eEvent)
         {
-            float clientProtocolVersion;
-            if (float.TryParse(eEvent.DataToString(), out clientProtocolVersion))
+            int clientProtocolVersion;
+            if (int.TryParse(eEvent.DataToString(), out clientProtocolVersion))
             {
                 var client = Authenticator.Client(eEvent.ClientId);
                 if (client != null)
